Migrate PageHeader to TS
